+++ eoreader/products/sar/s1_product.py
@@ -203,12 +203,8 @@ class S1Product(SarProduct):
         # Pre init done by the super class
         super()._pre_init(**kwargs)
 
-        # Check if COG in name
-        if (
-            "_COG" in self.filename
-            and self._need_snap
-            and not self._has_snap_10_or_higher()
-        ):
+        # Check if COG in name (S1 needs always SNAP)
+        if "_COG" in self.filename and not self._has_snap_10_or_higher():
             raise NotImplementedError(
                 "S1 COG products are only handled by SNAP 10.0 or higher. "
                 "Please upgrade your software to process this product."
@@ -277,7 +273,7 @@ class S1Product(SarProduct):
         self.product_type = S1ProductType.from_value(prod_type)
 
         if self.product_type == S1ProductType.GRD:
-            self.sar_prod_type = SarProductType.GDRG
+            self.sar_prod_type = SarProductType.GRD
         elif self.product_type == S1ProductType.SLC:
             self.sar_prod_type = SarProductType.CPLX
         else: